[Minor] Fixed a stability bug of the page history test.
The problem was due to the fact that if pages are stored too fast, they will have the same timestamp in the modification list (the granularity is 1 sec.) and they will be returned in an arbitrary order (depending on the server side query result).

This means that the modified page in the test, even if it's present in the modification list, could not be returned as the first item of that list, causing the assertion to fail. 

The solution was to wait 2 seconds before storing the page. This will guarantee that the timestamp will surely be different from any previous modification (and that the page will be on the top of the list if things work well)

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/enterprise/trunk@15898 f329d543-caf0-0310-9063-dda96c69346f